Fix a bug with Expander and React lifecycles
Pushing to ‘setActives’ when the component mounts was causing some
strange behaviour. Sometimes the expanders would get stuck after being
opened and not close again.

This fixes it by tracking the sticky and non-sticky ‘setActive’
functions separately which means we don’t need useEffect because we just
override the old one if the component re-renders.