Move images as soon as they’re picked
The ImageInput component contains the state of
which images have been picked. If we move images,
we need this to be reflected in this state or the
images will be broken.

Therefore, I think it makes sense to move images
as soon as they’re picked rather than doing this
‘at a distance’ in the answerQuestion workflow
which makes it harder to maintain consistency.