Set the Google Play Store upload channel to 'production'
At the moment, we have to promote releases to production manually
through the Google Play Console. However, because we're using Expo's
'over the air' update feature, we're effectively releasing as soon as
the bundles are uploaded to GitHub.

It may be that we want to use release channels for internal testing, but
this will at least keep the play store version and bundle version
consistent in the meantime.