Use project-specific translations if available
The user_interface_text field in /my_data can now include a list of
projects for which the translation applies. These translations appear
first in the list, followed by the defaults.

When pressing on a project on the Home screen, set the project and unset
it when going back from the Project screen.