Fix the hardware back button behaviour
When navigating to a new screen in react-navigation, the previous one
isn’t unmounted which means the registered BackHandler even listener
remains. Therefore, check if the screen also has focus before handling
the custom back button behaviour.

Based on this documentation:
https://reactnavigation.org/docs/4.x/custom-android-back-button-handling

In the latest version of react-navigation, there’s a useFocusEffect hook
so it might be worth upgrading to clean this code up a bit. See:
https://reactnavigation.org/docs/navigation-lifecycle#react-navigation-lifecycle-events